React router fix CRUD

+++ server/src/routes/blogRoutes.js
@@ -1,4 +1,5 @@
 const express = require("express");
+const { find } = require("../Model/BlogModel");
 const BlogModel = require("../Model/BlogModel");
 const UserModel = require("../Model/UserModel");
 
@@ -6,8 +7,8 @@ const router = express.Router();
 
 router.get("/all", (req, res) => {
   BlogModel.find()
-    .populate('categoryId', 'tag')
-    .populate('authorId', 'username')
+    .populate("categoryId", "tag")
+    .populate("authorId", "username")
     .then((data) => {
       res.send(data);
     })
@@ -18,8 +19,8 @@ router.get("/all", (req, res) => {
 
 router.get("/byid/:id", (req, res) => {
   BlogModel.findById(req.params.id)
-    .populate('categoryId', 'tag')
-    .populate('authorId', 'username')
+    .populate("categoryId", "tag")
+    .populate("authorId", "username")
     .then((blog) => {
       res.send(blog);
     })
@@ -29,7 +30,7 @@ router.get("/byid/:id", (req, res) => {
 });
 
 router.get("/bytitle/:title", (req, res) => {
-  BlogModel.find({title: req.params.title})
+  BlogModel.find({ title: req.params.title })
     .then((data) => {
       res.send(data);
     })
@@ -38,14 +39,28 @@ router.get("/bytitle/:title", (req, res) => {
     });
 });
 
+router.get("/myblog", (req, res) => { 
+  console.log(req.session.user.id)
+  BlogModel.find({ authorId: req.session.user.id})
+  .populate("categoryId", "tag")
+  .populate("authorId", "username")
+  .then((data) => {
+          res.send(data);
+        })
+        .catch(() => {
+          res.status(500).send("Unable to query Blogs!");
+        });
+      })
+
+
 router.post("/new-blog", (req, res) => {
   console.log(req.body);
-  let blogStructure ={
+  let blogStructure = {
     title: req.body.title,
     content: req.body.content,
-    categoryId:req.body.categoryId,
-    authorId:req.session.user.id
-  }
+    categoryId: req.body.categoryId,
+    authorId: req.session.user.id,
+  };
   BlogModel.create(blogStructure)
     .then((data) => {
       res.send(data);
@@ -61,29 +76,41 @@ router.patch("/update/:id", (req, res) => {
   const update = req.body;
   BlogModel.findById(id).then((blog) => {
     console.log(blog);
-    console.log(req.session.user)
-    if(blog.authorId == req.session.user.id){
+    console.log(req.session.user);
+    if (blog.authorId == req.session.user.id) {
       BlogModel.findByIdAndUpdate(id, update, {
         new: true,
         upsert: true,
-      }).then((data) => res.send(`${data} updated succesfully`));
-    }else{
+      }).then((data) => res.send(`${data.title} updated succesfully`));
+    } else {
       res.status(403).send("Unauthorised author, unable to change blog!");
     }
-  })
+  });
 });
 
 router.delete("/delete/:id", (req, res) => {
   let id = req.params.id;
-  BlogModel.findByIdAndDelete(id)
-    .then(() => {
-      res.send("data has been deleted succesfully");
-    })
-    .catch((err) => {
-      console.log(err);
-      res.status(500).send("unable to delete data");
-    });
+  console.log(id)
+  console.log(req.session.user);
+  BlogModel.findById(id).then((blogRecord)=>{
+
+    console.log('authorId',blogRecord.authorId)
+    console.log('sessionid', req.session.user.id)
+    if(blogRecord && blogRecord.authorId == req.session.user.id){
+      BlogModel.findByIdAndDelete(id)
+        .then(() => {
+          res.send("data has been deleted succesfully");
+        })
+    
+        .catch((err) => {
+          console.log(err);
+          res.status(400).send("unable to delete data");
+        });
+
+    }else{
+      res.status(403).send('unauthorised')
+    }
+  })
 });
 
 module.exports = router;
-
